1. updated to compile with Eclipse Indigo SR2 2. updated release.sh to work with Windows 7/cygwin 3. By default, prohibit unrestricted query permissions to everyone 4. To toglle 3), in config/qrcsp-config.xml, the parameter "UnrestrictedQueryPermissionsToAETitles" should be set to "ANY" / "NONE"

+++ release.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
+set -x
+umask 0007
+WIN=/cygdrive/c/src/eclipse/GatewaySetup/win; mkdir -p $WIN
+LIN=/cygdrive/c/src/eclipse/GatewaySetup/linux; mkdir -p $LIN
+MACOS=/cygdrive/c/src/eclipse/GatewaySetup/MacOS/Gateway-GUI.app/Contents/MacOS/Gateway; mkdir -p $MACOS
+SRC=/cygdrive/c/src/eclipse/gatewaye
+SRC_SETUP=/cygdrive/c/src/eclipse/GatewaySetup
 
-WIN=/cygdrive/c/Workspace/GatewaySetup/win
-LIN=/cygdrive/c/Workspace/GatewaySetup/linux
-MACOS=/cygdrive/c/Workspace/GatewaySetup/MacOS/Gateway-GUI.app/Contents/MacOS/Gateway
-SRC=/cygdrive/c/Workspace/xnd_stable/gatewaye
 
 cp -rf gatewaye.jar README.TXT $WIN/dist
 cp -rf config $WIN
@@ -17,14 +20,28 @@ cp -rf gatewaye.jar README.TXT $MACOS/dist
 cp -rf config $MACOS
 cp -rf lib $MACOS/dist
 
+
+read -p "continue (y/n)?" status
+if [ "$status" == "n" ]; then exit 1; fi
+
 ver=`date +"%b%Y"`
 ISETUP="/cygdrive/c/Program Files (x86)/Inno Setup 5/ISCC.exe"
-pushd /cygdrive/c/Workspace/GatewaySetup/
+pushd $SRC_SETUP
+chmod -R 777 *
 "$ISETUP" gateway.iss
+chmod -R 777 *
 "$ISETUP" gateway-bundled.iss
+chmod -R 777 *
+
+read -p "continue (y/n)?" status
+if [ "$status" == "n" ]; then exit 1; fi
 
 T=/cygdrive/y/ftp/pub/xnd/download/gw
 mv -f $T/XNAT_Gateway_* $T/old
+
+read -p "continue (y/n)?" status
+if [ "$status" == "n" ]; then exit 1; fi
+
 
 pushd linux
 rm *.zip
@@ -34,6 +51,10 @@ chmod o+rwx XNAT_Gateway_Linux_${ver}.zip
 cp -f XNAT_Gateway_Linux_${ver}.zip $T
 popd 
 
+read -p "continue (y/n)?" status
+if [ "$status" == "n" ]; then exit 1; fi
+
+
 pushd MacOS
 rm *.zip
 chmod -R o+rwx *
@@ -41,6 +62,9 @@ zip -r XNAT_Gateway_MacOS_${ver}.zip *
 chmod o+rwx XNAT_Gateway_MacOS_${ver}.zip
 cp -f XNAT_Gateway_MacOS_${ver}.zip $T
 popd
+
+read -p "continue (y/n)?" status
+if [ "$status" == "n" ]; then exit 1; fi
 
 pushd win/Output
 chmod -R o+rwx *
